Remove unused custom support
If a user wants to support other codecs
they can either add it to this library, or just
discriminate in their own code between
codecs supplied by smush, and the other ones.

This also makes the Quality and Encoding
types Clone, which cleans up the API usage a bit
for example in the example program.